Extract cursor_shape.c from misc2.c and types/consts from structs.h
