+++ src/misc2.h
@@ -65,9 +65,6 @@ int after_pathsep(char_u *b, char_u *p);
 int same_directory(char_u *f1, char_u *f2);
 int vim_chdirfile(char_u *fname);
 int illegal_slash(char *name);
-char_u *parse_shape_opt(int what);
-int get_shape_idx(int mouse);
-void update_mouseshape(int shape_idx);
 int vim_chdir(char_u *new_dir);
 int get_user_name(char_u *buf, int len);
 void sort_strings(char_u **files, int count);
